core/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
billiard==4.2.1
black==24.8.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==4.2.18
django-celery-beat==2.7.0
django-ckeditor==6.7.2
django-filter==24.3
django-js-asset==2.2.0
django-mail-templated==2.6.5
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
Faker==35.2.2
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itsdangerous==2.2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
Markdown==3.7
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg-binary==3.2.5
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
